+++ NEWS.md
@@ -1,5 +1,7 @@
 # era (development version)
 
+* Added yr_sort() for chronological ordering of year vectors (#44)
+
 # era 0.4.1
 
 This patch fixes a malfunctioning test that was causing CRAN checks to fail. There are no significant changes.